+++ src/main/java/com/sql/sqlextra/entity/EmailVisit.java
@@ -17,7 +17,7 @@ public class EmailVisit {
     private Long id;
 
     @Column(name = "id_account", nullable = false)
-    private Long idAccount;
+    private Integer idAccount;
 
     @Column(name = "visit_date")
     private Integer visitDate;
